Moved canCoder and canCoder func to ArmIOController and added canCoder Rotation to Arm inputs
