+++ src/main/java/frc/robot/subsystems/arm/Arm.java
@@ -16,8 +16,6 @@ public class Arm extends SubsystemBase {
     private ArmIO io;
     private final ArmIOInputsAutoLogged inputs = new ArmIOInputsAutoLogged();
     private boolean enabled;
-    private CANcoder canCoder;
-
 
     public Arm(boolean enabled, ArmIO io) {
         if (enabled) {
@@ -25,12 +23,6 @@ public class Arm extends SubsystemBase {
             io.setup();
         }
         this.enabled = enabled;
-
-        canCoder = new CANcoder(Constants.kArmCanCoderID);
-        CANcoderConfiguration config = new CANcoderConfiguration();
-        config.MagnetSensor.MagnetOffset = Constants.kArmCanCoderOffset;
-        config.MagnetSensor.SensorDirection = kArmCanCoderReversed;
-        canCoder.getConfigurator().apply(config);
     }
 
     @Override
@@ -44,10 +36,6 @@ public class Arm extends SubsystemBase {
         Logger.processInputs("Arm", inputs);
     }
 
-    public boolean isCoralInside() {
-        return RobotState.isCoralInArm();
-    }
-
     //--Commands
 
     public Command setAngle(Rotation2d angle){
@@ -57,13 +45,6 @@ public class Arm extends SubsystemBase {
         return Commands.runOnce(() -> {io.getController().setPosition(angle.getRadians());});
     }
 
-    public Rotation2d getAngle(){
-        if (!enabled) {
-            return Rotation2d.kZero;
-        }
-        return Rotation2d.fromRadians(io.getController().getPosition());
-    }
-
     public boolean atGoal(){
         if (!enabled){
             return true;
@@ -71,21 +52,17 @@ public class Arm extends SubsystemBase {
         return io.getController().atGoal();
     }
 
-    public Rotation2d getCANCoder(){
-        return Rotation2d.fromRotations(canCoder.getAbsolutePosition().getValueAsDouble());
-    }
-
     public Command reset(){
         if (!enabled){
             return Commands.none();
         }
-        return Commands.runOnce(() -> {io.getController().setEncoder(getCANCoder().getRadians());}).andThen(setAngle(Rotation2d.fromDegrees(Constants.ArmPositions.Close.get())));
+        return Commands.runOnce(() -> {io.getController().setEncoder(io.getCANCoder().getRadians());}).andThen(setAngle(Rotation2d.fromDegrees(Constants.ArmPositions.Close.get())));
     }
 
     public boolean isReset(){
        if (!enabled){
             return true;
        }
-       return Math.abs(getCANCoder().getDegrees() - Constants.ArmPositions.Close.get()) < Constants.kArmControllerConstants.real.positionGoalTolerance;
+       return Math.abs(io.getCANCoder().getDegrees() - Constants.ArmPositions.Close.get()) < Constants.kArmControllerConstants.real.positionGoalTolerance;
     }
 }
