+++ src/main/java/frc/robot/subsystems/arm/ArmIO.java
@@ -1,14 +1,20 @@
 package frc.robot.subsystems.arm;
 
+import edu.wpi.first.math.geometry.Rotation2d;
 import frc.lib.NinjasLib.controllers.Controller;
 import org.littletonrobotics.junction.AutoLog;
 
 public interface ArmIO {
     @AutoLog
     class ArmIOInputs extends Controller.ControllerIOInputs {
+        Rotation2d AbsoluteAngle;
     }
 
     default void setup() {
+    }
+
+    default Rotation2d getCANCoder(){
+        return null;
     }
 
     default Controller getController() {
